Initial implementation. Tests not yet passing.
Looks like we're running into issues with const generics, or the idea
that we are passing in an array of a certain type with a fixed length
that we don't yet know. In this case, the fixed length is the number of
taps, which is defined by the length of the coefficient array passed in
during the creation of the filter.

I've tried working around this issue with the use of `ExactSizeIterator`
and `IndexMut` traits, but these are not implemented for integer arrays
(according to the compiler errors). Need to find another approach.